Move things around and run prettier
