+++ app/routes/_docs/nav/NavExternalLink.fragment.gql
